# backend/openapi/agro-trade.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/onboarding/seller": {
      "post": {
        "operationId": "OnboardingController_completeSellerOnboarding",
        "summary": "Complete seller onboarding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellerOnboardingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingUserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/onboarding/buyer": {
      "post": {
        "operationId": "OnboardingController_completeBuyerOnboarding",
        "summary": "Complete buyer onboarding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerOnboardingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingUserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/onboarding/transporter": {
      "post": {
        "operationId": "OnboardingController_completeTransporterOnboarding",
        "summary": "Complete transporter onboarding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransporterOnboardingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingUserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/onboarding/status": {
      "get": {
        "operationId": "OnboardingController_getOnboardingStatus",
        "summary": "Get onboarding status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingStatusResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Onboarding"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Agro-Trade API",
    "description": "API documentation for Agro-Trade platform",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SellerOnboardingDto": {
        "type": "object",
        "properties": {}
      },
      "OnboardingUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "object"
          },
          "phoneNumber": {
            "type": "object"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "FARMER",
              "BUYER",
              "TRANSPORTER",
              "COMPANY_ADMIN",
              "INSPECTOR"
            ]
          },
          "onboardingCompleted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "onboardingCompleted"
        ]
      },
      "BuyerOnboardingDto": {
        "type": "object",
        "properties": {}
      },
      "TransporterOnboardingDto": {
        "type": "object",
        "properties": {}
      },
      "OnboardingCompanyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "legalName": {
            "type": "object"
          },
          "registrationNumber": {
            "type": "object"
          },
          "vatNumber": {
            "type": "object"
          },
          "phoneNumber": {
            "type": "object"
          },
          "email": {
            "type": "object"
          },
          "website": {
            "type": "object"
          }
        },
        "required": [
          "id"
        ]
      },
      "OnboardingAddressDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "addressType": {
            "type": "string",
            "enum": [
              "BILLING",
              "SHIPPING",
              "FARM",
              "WAREHOUSE",
              "OFFICE",
              "PICKUP",
              "DELIVERY",
              "OTHER"
            ]
          },
          "label": {
            "type": "object"
          },
          "street": {
            "type": "object"
          },
          "cityId": {
            "type": "object"
          },
          "postalCode": {
            "type": "object"
          },
          "country": {
            "type": "object"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "isDefault": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "addressType"
        ]
      },
      "OnboardingTruckDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "plateNumber": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "FLATBED",
              "REFRIGERATED",
              "TANKER",
              "CONTAINER",
              "CURTAIN_SIDE",
              "BOX_TRUCK",
              "OTHER"
            ]
          },
          "currentLocation": {
            "type": "object"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "isAvailable": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "plateNumber",
          "type",
          "isAvailable"
        ]
      },
      "OnboardingDataDto": {
        "type": "object",
        "properties": {
          "company": {
            "$ref": "#/components/schemas/OnboardingCompanyDto"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OnboardingAddressDto"
            }
          },
          "trucks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OnboardingTruckDto"
            }
          }
        }
      },
      "OnboardingStatusResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "FARMER",
              "BUYER",
              "TRANSPORTER",
              "COMPANY_ADMIN",
              "INSPECTOR"
            ]
          },
          "isComplete": {
            "type": "boolean"
          },
          "onboardingCompleted": {
            "type": "boolean"
          },
          "missingFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "data": {
            "$ref": "#/components/schemas/OnboardingDataDto"
          }
        },
        "required": [
          "userId",
          "role",
          "isComplete",
          "onboardingCompleted",
          "missingFields",
          "data"
        ]
      },
...
    }
  }
}